+++ common/src/main/java/net/satisfy/beachparty/core/block/LineConnectingBlock.java
@@ -23,8 +23,6 @@ import org.jetbrains.annotations.Nullable;
 
 @SuppressWarnings("deprecation")
 public class LineConnectingBlock extends Block {
-
-
     public static final DirectionProperty FACING;
     public static final EnumProperty<BeachpartyUtil.LineConnectingType> TYPE;