+++ common/src/main/java/net/satisfy/beachparty/core/event/CommonEvents.java
@@ -6,8 +6,11 @@ import dev.architectury.event.events.common.LootEvent;
 import dev.architectury.event.events.common.PlayerEvent;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.DamageTypes;
@@ -28,6 +31,8 @@ import net.satisfy.beachparty.core.block.BeachParasolBlock;
 import net.satisfy.beachparty.core.block.RadioBlock;
 import net.satisfy.beachparty.core.registry.ObjectRegistry;
 import org.jetbrains.annotations.Nullable;
+
+import static net.satisfy.beachparty.core.block.RadioBlock.syncRadios;
 
 public class CommonEvents {
 
@@ -56,13 +61,11 @@ public class CommonEvents {
         ItemStack itemInHand = player.getItemInHand(hand);
 
         if (itemInHand.getItem() == ObjectRegistry.POOL_NOODLE.get()) {
-            Vec3 knockbackDirection = new Vec3(
-                    entity.getX() - player.getX(),
-                    0.2,
-                    entity.getZ() - player.getZ()
-            ).normalize().scale(1.5);
+            Vec3 knockbackDirection = new Vec3(entity.getX() - player.getX(), 0.6, entity.getZ() - player.getZ()).normalize().scale(1.5);
 
             entity.push(knockbackDirection.x, knockbackDirection.y, knockbackDirection.z);
+
+            level.playSound(null, entity.getX(), entity.getY(), entity.getZ(), SoundEvents.PARROT_IMITATE_SLIME, SoundSource.PLAYERS, 1.0F, 1.5F);
 
             return EventResult.interruptTrue();
         }
@@ -129,4 +132,15 @@ public class CommonEvents {
             return LootTableReference.lootTableReference(table);
         }
     }
+
+    public static void onServerStarted(MinecraftServer server) {
+        for (ServerLevel level : server.getAllLevels()) {
+            syncRadios(level);
+        }
+    }
+
+    public static void onPlayerJoin(ServerPlayer player) {
+        syncRadios(player.getLevel());
+    }
+
 }