+++ common/src/main/java/net/satisfy/beachparty/core/item/PoolNoodleItem.java
@@ -18,12 +18,12 @@ public class PoolNoodleItem extends SwordItem implements DyeableLeatherItem {
 
     private final Multimap<Attribute, AttributeModifier> attributeModifiers;
 
-    public PoolNoodleItem(Tier toolMaterial, int attackDamage, float attackSpeed, Properties properties) {
-        super(toolMaterial, attackDamage, attackSpeed, properties);
+    public PoolNoodleItem(Properties properties) {
+        super(Tiers.WOOD, 0, -1.4F, properties);
 
         ImmutableMultimap.Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
-        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Weapon modifier", attackDamage, AttributeModifier.Operation.ADDITION));
-        builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Weapon modifier", attackSpeed, AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Weapon modifier", 0, AttributeModifier.Operation.ADDITION));
+        builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Weapon modifier", -1.4F, AttributeModifier.Operation.ADDITION));
         builder.put(Attributes.ATTACK_KNOCKBACK, new AttributeModifier("Weapon knockback", 3, AttributeModifier.Operation.ADDITION));
 
         this.attributeModifiers = builder.build();
@@ -33,7 +33,6 @@ public class PoolNoodleItem extends SwordItem implements DyeableLeatherItem {
     public @NotNull Multimap<Attribute, AttributeModifier> getDefaultAttributeModifiers(@NotNull EquipmentSlot slot) {
         return slot == EquipmentSlot.MAINHAND ? this.attributeModifiers : super.getDefaultAttributeModifiers(slot);
     }
-
 
     @Override
     public void appendHoverText(@NotNull ItemStack itemStack, Level world, List<Component> tooltip, @NotNull TooltipFlag tooltipContext) {